+++ CoreWorker.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-CoreWorker::CoreWorker(int id, float delayPerExec, int quantumSlice) 
+CoreWorker::CoreWorker(int id, float delayPerExec, float quantumSlice) 
     : id(id), delayPerExec(delayPerExec), quantumSlice(quantumSlice), running(false), processAssigned(false) {}
 
 CoreWorker::~CoreWorker() {
@@ -24,16 +24,22 @@ void CoreWorker::setProcess(std::shared_ptr<Process> process) {
 }
 
 void CoreWorker::runProcess() {
+
+    // For SJF and FCFS
     if (quantumSlice == 0) {
+        // While there is still a process, run
         while (currentProcess && !currentProcess->isFinished()) {
             currentProcess->execute();
             std::this_thread::sleep_for(std::chrono::duration<float>(delayPerExec));
         }
 
+        // Process is finished, notify Scheduler
         if (currentProcess && currentProcess->isFinished()) {
             finishProcess();
         }
     }
+
+    // For RR
     else {
         for (int i = 0; i < quantumSlice; i++) {
             if (currentProcess->isFinished()) {
@@ -58,18 +64,18 @@ void CoreWorker::runProcess() {
 
 void CoreWorker::finishProcess() {
     std::lock_guard<std::mutex> lock(coreMutex);
-    currentProcess.reset();
-    processAssigned = false;
+    currentProcess.reset(); // Reset the current process
+    processAssigned = false; // No more process assigned
 }
 
 bool CoreWorker::isAvailable() {
     std::lock_guard<std::mutex> lock(coreMutex);
-    return !processAssigned;
+    return !processAssigned; // Core is available if no process is assigned
 }
 
 bool CoreWorker::isAssignedProcess() {
 	std::lock_guard<std::mutex> lock(coreMutex);
-	return processAssigned;
+	return processAssigned; // Has a process assigned
 }
 
 int CoreWorker::getID() {
@@ -78,7 +84,7 @@ int CoreWorker::getID() {
 
 void CoreWorker::start() {
     running = true;
-    coreThread = std::thread(&CoreWorker::run, this);
+    coreThread = std::thread(&CoreWorker::run, this); // Start the core thread
 }
 
 void CoreWorker::stop() {
@@ -93,6 +99,7 @@ void CoreWorker::stop() {
 }
 
 void CoreWorker::run() {
+    // Core thread loop
     while (true) {
         {
             std::lock_guard<std::mutex> lock(coreMutex);
@@ -107,6 +114,7 @@ void CoreWorker::run() {
     }
 }
 
+// Set the process completion callback
 void CoreWorker::setProcessCompletionCallback(std::function<void(std::shared_ptr<Process>)> callback) {
     std::lock_guard<std::mutex> lock(coreMutex);
     processCompletionCallback = callback;
